Initialize querydata latlon-cache after loading the data
If the latlon-cache is not initialized, requests using the model will race to be
the first to initialize it, and the others will wait in the queue. Since access
to metadata is controlled using the same mutex as querydata updates, metadata
requests may be exceedingly slow if they have to initialize the latlon caches too.
If a write lock is requested while the metadata is still being processed, all
subsequent requests requiring read-only access will be blocked.

Other possible improvements to consider:
- use timed write locks not to block read requests for too long
- avoid long duration lock while requesting metadata
- avoid latlon-cache initialization if the data is too old for the repository