Optimized loading algorithm to use an internal coordinate cache for NFmiQueryData objects
The engine now maintains an internal cache of the LatLonCache objects
used by NFmiQueryData objects. NFmiQueryData now supports calculating
a hash value for the grid. If the engine cache contains a LatLonCache
object for the NFmiQueryData object based on the hash value for the
grid, we simply set the cache externally into the NFmiQueryData object
instead of letting it calculate by itself when the first latlon
request is performed. On the other hand, if no cached coordinates
exist, we request the NFmiQueryData object calculate the coordinates
and then request them for caching. The coordinates are stored both in
NFmiQueryData and in the cache as shared pointers to minimize memory
use.