Added setting valid_points_cache_dir with default value "/var/smartmet/querydata/validpoints"
querydata-engine will establish which grid points have valid data if fullgrid=true. This can
be very slow for high resolution grids. The code will now serialize the results to a cache
so that a possible restart can be much faster. Cached files will be deleted as the models
are removed from memory. However, this does not guarantee all cache files will be deleted,
an extra cron-routine is needed to delete very old files which have not been accessed
in a long time.

A restart may now take less than 10 seconds instead of several minutes.