Include configurable model properties into the model hash value
We may shortly change some models from using relative U/V wind components
to absolute ones. The modification has an effect on products, and hence
must be included in the hash value for the model. Climatology and full-grid
information is now also included, even though that information is very
unlikely to change.