Default number_to_keep is now 2 instead of 1
Imagine a refresh interval of 60 seconds. Different backends will see
new files at different times. When new data arrives, there is a one
minute interval when any one of the backends may still only see the
older files. If a qengine status query arrives during this period,
an intersection of the common files is taken. If number_to_keep is
only one, this intersection will be empty during the interval
when the first backend has noticed the new data and the last one
has noticed it too. This may cause problems for production which
assumes there is always common content available.

If the default value is changed to two, this problem will not
occur, since the second oldest file will be common to all backends
during the one minute refresh interval.